# apps/backend-go/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  models.APIResponse:
    properties:
      count:
        type: integer
      data: {}
      error:
        type: string
      has_more:
        type: boolean
      next_cursor:
        type: string
      success:
        type: boolean
    type: object
...
info:
  contact:
    name: gomo6 Team
    url: https://gomo6.wtf
  description: gomo6 — open social platform API. Create bots, integrations, and apps.
  license:
    name: MIT
  termsOfService: https://gomo6.wtf/terms
  title: gomo6 API
  version: "1.0"
paths:
...
  /user/drops:
    get:
      description: Get the authenticated user's drops balance
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.APIResponse'
      security:
      - BearerAuth: []
      summary: Get drops balance
      tags:
      - Drops
...
schemes:
- https
- http
securityDefinitions:
  BearerAuth:
    description: Bearer token for API authentication
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"